[UIK-5765][chore] fix lint issues/ts projects refactoring
