+++ package.json
@@ -4,7 +4,7 @@
     "dev": "pnpm storybook",
     "start": "pnpm storybook",
     "watch-semcore": "tsc -b semcore --emitDeclarationOnly --watch",
-    "build": "pnpm run clean && pnpm --filter @semcore/theme run build && pnpm --filter @semcore/core run build && pnpm build:icons && pnpm build:illustration && pnpm --filter @semcore/base-components run build && pnpm --filter @semcore/dropdown run build && pnpm --filter @semcore/* --filter !@semcore/icon --filter !@semcore/core --filter !@semcore/base-components --filter !@semcore/illustration --filter !@semcore/dropdown --filter !@semcore/ui run build",
+    "build": "pnpm run clean && pnpm -r --filter @semcore/* --filter !@semcore/ui build && pnpm --filter @semcore/ui generate-reexport",
     "process-theme": "pnpm --filter @semcore/theme build",
     "install-n-test": "pnpm install && pnpm test",
     "install-n-browser-test": "pnpm install && pnpm browser-test",
@@ -29,9 +29,11 @@
     "clean:illustration": "pnpm --filter @semcore/illustration run clean",
     "format": "biome format ./semcore ./tools ./website ./stories ./.ci --config-path . --write",
     "lint:es": "npx eslint",
-    "lint:ts": "NODE_OPTIONS=--max-old-space-size=8192 pnpm lint:ts:semcore && tsc --noEmit && pnpm lint:ts:website",
-    "lint:ts:semcore": "tsc -b semcore --emitDeclarationOnly --verbose",
-    "lint:ts:website": "cd website && tsc --noEmit",
+    "lint:ts": "pnpm lint:ts:stories && pnpm lint:ts:playground && pnpm lint:ts:tools && pnpm lint:ts:website",
+    "lint:ts:stories": "tsc --noEmit --project stories --diagnostics",
+    "lint:ts:playground": "tsc --noEmit --project playground --diagnostics",
+    "lint:ts:tools": "tsc --noEmit --project tools --diagnostics",
+    "lint:ts:website": "tsc --noEmit --project website --diagnostics",
     "lint:css": "stylelint **/*.css",
     "lint:docs": "vale README.md CONTRIBUTING.md semcore/*/README.md semcore/*/CHANGELOG.md tools/*/README.md tools/*/CHANGELOG.md website/docs",
     "prepare": "husky install",
