+++ stories/package.json
@@ -16,6 +16,9 @@
   "devDependencies": {
     "@types/d3-scale": "3.3.0",
     "@types/d3-shape": "3.1.7",
-    "@types/react-virtualized": "9.22.0"
+    "@types/react-virtualized": "9.22.0",
+    "@semcore/ui": "workspace:*",
+    "@semcore/illustration": "workspace:*",
+    "@semcore/icon": "workspace:*"
   }
 }
