+++ tools/@types/tsconfig.json
@@ -1,4 +1,4 @@
 {
-  "extends": "../../tsconfig.json",
+  "extends": "../tsconfig.json",
   "include": ["globals", "styles", "deps", "vite"]
 }
