--- tools/babel-plugin-react-semcore/tsconfig.json
@@ -1,4 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "include": ["src", "lib", "index.js"]
-}
