--- tools/babel-preset-ui/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "include": ["src/**/*.ts", "src/**/*.tsx", "src/**/*.json"],
-  "compilerOptions": {
-    "rootDir": "./src"
-  }
-}
