+++ tools/continuous-delivery/tsconfig.json
@@ -1,7 +1,3 @@
 {
-  "extends": "../../tsconfig.json",
-  "include": ["src/**/*.ts", "src/**/*.tsx", "src/**/*.json"],
-  "compilerOptions": {
-    "rootDir": "./src"
-  }
+  "extends": "../tsconfig.json",
 }
