+++ tools/esbuild-plugin-semcore/tsconfig.json
@@ -1,8 +1,4 @@
 {
-  "extends": "../../tsconfig.json",
-  "include": ["src/**/*.ts", "src/**/*.tsx", "src/**/*.json"],
-  "compilerOptions": {
-    "rootDir": "./src",
-    "allowImportingTsExtensions": true
-  }
+  "extends": "../tsconfig.json",
+  "include": ["src/**/*.ts"],
 }
