+++ tools/gcs-upload/tsconfig.json
@@ -1,4 +1,3 @@
 {
-  "extends": "../../tsconfig.json",
-  "include": ["index.ts", "bin"]
+  "extends": "../tsconfig.json",
 }
