+++ tools/i18n-unplugin/package.json
@@ -9,8 +9,7 @@
   "author": "UI-kit team <ui-kit-team@semrush.com>",
   "license": "MIT",
   "scripts": {
-    "build": "pnpm semcore-builder --source=ts",
-    "test": "echo \"no tests\""
+    "build": "pnpm semcore-builder"
   },
   "dependencies": {
     "unplugin": "1.2.0"
