+++ tools/theme/package.json
@@ -6,7 +6,7 @@
   "license": "MIT",
   "type": "module",
   "scripts": {
-    "build": "node --experimental-transform-types src/processor.ts && tsc --project tsconfig-theme.json",
+    "build": "node --experimental-transform-types src/processor.ts && tsc",
     "check": "node --experimental-transform-types src/checker.ts",
     "watch": "chokidar \"src/**/*.ts\" -c \"node --experimental-transform-types src/processor.ts --watch\""
   },
