--- tools/theme/tsconfig-theme.json
@@ -1,15 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "include": [
-    "src/theme.ts"
-  ],
-  "compilerOptions": {
-    "lib": [
-      "ESNext"
-    ],
-    "rootDir": "./src",
-    "allowImportingTsExtensions": true,
-    "emitDeclarationOnly": true,
-    "outDir": "./lib"
-  }
-}
