--- tsconfig.json
@@ -1,21 +1,0 @@
-{
-  "extends": "./tsconfig.base.json",
-  "compilerOptions": {
-    "allowImportingTsExtensions": true,
-    "paths": {
-      "@semcore/illustration/*": ["./semcore/illustration/lib/*"],
-      "@semcore/illustration": ["./semcore/illustration"],
-      "@semcore/icon/*": ["./semcore/icon/lib/*"],
-      "@semcore/ui/d3-chart": ["./semcore/d3-chart"],
-      "@semcore/ui/core/lib/*": ["./semcore/core/src/*"],
-      "@semcore/ui/*": ["./semcore/*/src"],
-      "@semcore/*": ["./semcore/*/src"],
-      "@tools/*": ["./tools/*"]
-    },
-  },
-  "include": ["tools", "stories", "playground"],
-  "exclude": [
-    "node_modules",
-    "tools/icon-transform-svg/__tests__"
-  ]
-}
